# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Machine Test
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /students:
    x-swagger-router-controller: user
    post:
      tags: ['Students']
      description: create student
      summary: create student
      # used as the method name of the controller
      operationId: createSrudent
      parameters:
        - name: body
          in: body
          description: input data
          schema:
           $ref: '#/definitions/createStudent'
          required: true
      responses:
        "201":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/response"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/response"
    get:
      tags: ['Students']
      summary: get all students
      description: get all students
      # used as the method name of the controller
      operationId: getAllStudents
      parameters:
        - name: page
          in: query
          description: page no
          required: false
          type: number
          default: 1
        - name: limit
          in: query
          description: per page
          required: false
          type: number
          default: 10
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/response"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/response"
  /students/{id}:
    x-swagger-router-controller: user
    patch:
      tags: ['Students']
      description: update student
      summary: update student
      # used as the method name of the controller
      operationId: updateStudent
      parameters:
        - name: id
          in: path
          description:  student id
          type: number
          required: true
        - name: body
          in: body
          description: input data
          schema:
            $ref: '#/definitions/createStudent'
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/response"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/response"
    delete:
      tags: ['Students']
      summary: delete student
      description: delete student
      # used as the method name of the controller
      operationId: deleteStudent
      parameters:
        - name: id
          in: path
          description:  record id
          type: string
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/response"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/response"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  response:
    required:
      - message
    properties:
      message:
        type: string
  createStudent:
    properties:
      name:
        type: string
      marks:
        type: number
      percentage:
        type: number
      subjects:
        type: array
        items:
          type: object
          properties:
            name: string